migrated forms menu to forms.php
